Parte da manhã dia 10.03

+++ src/main/operacao/Operacao.java
@@ -14,7 +14,7 @@ public class Operacao {
 	int pdv = -1;
 	int[] denominacao = new int[6];
 	int total;
-	double valorAB;
+	int valorAB;
 	Long codigo_lacre = (long) -1;
 	
 
@@ -22,7 +22,7 @@ public class Operacao {
 	public double getValorAB() {
 		return valorAB;
 	}
-	public void setValorAB(double valorAB) {
+	public void setValorAB(int valorAB) {
 		this.valorAB = valorAB;
 	}
